+++ NEWS.md
@@ -1,4 +1,43 @@
-# Seurat 4.0.2 (2020-03-20)
+## Unreleased
+## Changes
+- Update documentation for `to.upper` parameter in `Load10X_Spatial()` ([#4576](https://github.com/satijalab/seurat/issues/4576))
+- Update concept tags for `RunSPCA()` ([#4978](https://github.com/satijalab/seurat/discussions/4987))
+
+## Seurat 4.0.4 (2020-08-19)
+## Added
+- Add `reduction` parameter to `BuildClusterTree()` ([#4598](https://github.com/satijalab/seurat/issues/4598))
+- Add DensMAP option to `RunUMAP()` ([#4630](https://github.com/satijalab/seurat/pull/4630))
+- Add `image` parameter to `Load10X_Spatial()` and `image.name` parameter to `Read10X_Image()` ([#4641](https://github.com/satijalab/seurat/pull/4641))
+- Add `ReadSTARsolo()` function to read output from STARsolo
+- Add `densify` parameter to `FindMarkers()`
+- Add `ReadParsebio()` function to read output from Parse Biosciences
+- Add the `image.alpha` parameter to `SpatialDimPlot()` and `SpatialFeaturePlot()`
+- Add support for the correlation metric in `RunUMAP` ([#4972](https://github.com/satijalab/seurat/issues/4972))
+
+## Changes
+- Warn and continue rather than erroring if not all features are available in `FindSpatiallyVariableFeatures()` ([#4611](https://github.com/satijalab/seurat/issues/4611))
+- Bug fix for SCT-based integration in selecting proper reference model ([#4355](https://github.com/satijalab/seurat/issues/4355))
+- Bug fix for reading from absolute paths in ReadMtx ([#4723](https://github.com/satijalab/seurat/issues/4723)) 
+- Bug fix in SingleCellExperiment conversion ([#4633](https://github.com/satijalab/seurat/issues/4633))
+- Bug fix in `FindVariableFeatures()` when using `selection.method = "mvp"` and `binning.method = "equal_frequency"` ([#4712](https://github.com/satijalab/seurat/issues/4712))
+- Bug fix in `DoHeatmap()` to remove random characters from plot legend([#4660](https://github.com/satijalab/seurat/issues/4660))
+- Fix cell renaming in `RunCCA()`
+- Fix issue in SingleCellExperiment conversion where the mainExp would not be set properly
+- Fix for default dispersion info displayed in `VariableFeaturePlot()`
+
+## Seurat 4.0.3 (2020-06-10)
+## Added
+- Add `jitter` parameter to `FeatureScatter()` 
+
+## Changes
+- Fix issues with `as.SingleCellExperiment.Seurat()` for the latest verion of SingleCellExperiment ([#4532](https://github.com/satijalab/seurat/pull/4532))
+- Ensure proper reference.reduction is used in `MapQuery()`
+- Fix to `UpdateSymbolList()`, no longer searches aliases and exposes the `search.types` parameter in `GeneSymbolThesarus()` ([#4545](https://github.com/satijalab/seurat/issues/4545))
+- Transfer `scale.data` slot as well when converting with `as.SingleCellExperiment.Seurat()`
+- Enable `alpha` parameter for `SpatialDimPlot()`
+- Fix `as.SingleCellExperiment.Seurat()` conversion for atypical `reducedDim` components
+
+# Seurat 4.0.2 (2020-05-20)
 ## Added
 - New `AddAzimuthScores()` and `AddAzimuthResults()` functions
 - Add `shuffle` parameter to `FeatureScatter()` ([#4280](https://github.com/satijalab/seurat/pull/4280))